体重の一週間前比を表示する機能

+++ backend/database/migrations/2021_01_24_143101_create_weights_table.php
@@ -17,8 +17,8 @@ class CreateWeightsTable extends Migration
             $table->bigIncrements('id');
             $table->unsignedBigInteger('user_id');
             $table->timestamps();
-            $table->integer('weight')->nullable();
-            
+            //$table->integer('weight')->nullable();
+            $table->float('weight', 3, 1)->nullable();
             // 外部キー制約
             $table->foreign('user_id')->references('id')->on('users');
         });
